rework tournament setup
Combine the different tracks into a single tournament format, where
teams may qualify for the main track or stay in a challenge track.

Also, switch the sectioning depth to unify everything related to a
tournament under a single section.

Lastly, adapt the round-robin and playoff phase of the main track, such
that both are evaluated based on the same scoring scheme and allow for
bo3/bo5 games during playoff